PRSD-891: Move property deregistration to propertyDeregistrationService
