+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/forms/journeys/factories/PropertyDeregistrationJourneyFactory.kt
@@ -6,8 +6,8 @@ import uk.gov.communities.prsdb.webapp.controllers.DeregisterPropertyController
 import uk.gov.communities.prsdb.webapp.forms.journeys.PropertyDeregistrationJourney
 import uk.gov.communities.prsdb.webapp.models.viewModels.emailModels.PropertyDeregistrationConfirmationEmail
 import uk.gov.communities.prsdb.webapp.services.EmailNotificationService
+import uk.gov.communities.prsdb.webapp.services.PropertyDeregistrationService
 import uk.gov.communities.prsdb.webapp.services.PropertyOwnershipService
-import uk.gov.communities.prsdb.webapp.services.PropertyRegistrationService
 import uk.gov.communities.prsdb.webapp.services.factories.JourneyDataServiceFactory
 
 @Component
@@ -15,7 +15,7 @@ class PropertyDeregistrationJourneyFactory(
     private val validator: Validator,
     private val journeyDataServiceFactory: JourneyDataServiceFactory,
     private val propertyOwnershipService: PropertyOwnershipService,
-    private val propertyRegistrationService: PropertyRegistrationService,
+    private val propertyDeregistrationService: PropertyDeregistrationService,
     private val confirmationEmailSender: EmailNotificationService<PropertyDeregistrationConfirmationEmail>,
 ) {
     fun create(propertyOwnershipId: Long) =
@@ -23,7 +23,7 @@ class PropertyDeregistrationJourneyFactory(
             validator,
             journeyDataServiceFactory.create(DeregisterPropertyController.getPropertyDeregistrationPath(propertyOwnershipId)),
             propertyOwnershipService,
-            propertyRegistrationService,
+            propertyDeregistrationService,
             confirmationEmailSender,
             propertyOwnershipId,
         )
