+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/services/LandlordDeregistrationService.kt
@@ -15,4 +15,13 @@ class LandlordDeregistrationService(
         landlordRepository.deleteByBaseUser_Id(baseUserId)
         oneLoginUserRepository.deleteIfNotLocalAuthorityUser(baseUserId)
     }
+
+    @Transactional
+    fun deregisterLandlordsProperties(baseUserId: String) {
+        // retrieve all property ownerships
+
+        // delete all licenses
+        // delete all properties
+        // delete all property ownerships
+    }
 }
