+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/controllers/DeregisterPropertyControllerTests.kt
@@ -8,8 +8,8 @@ import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
-import org.springframework.boot.test.mock.mockito.MockBean
 import org.springframework.security.test.context.support.WithMockUser
+import org.springframework.test.context.bean.override.mockito.MockitoBean
 import org.springframework.test.web.servlet.get
 import org.springframework.web.context.WebApplicationContext
 import uk.gov.communities.prsdb.webapp.constants.CONFIRMATION_PATH_SEGMENT
@@ -18,8 +18,8 @@ import uk.gov.communities.prsdb.webapp.controllers.DeregisterPropertyController.
 import uk.gov.communities.prsdb.webapp.forms.journeys.PropertyDeregistrationJourney
 import uk.gov.communities.prsdb.webapp.forms.journeys.factories.PropertyDeregistrationJourneyFactory
 import uk.gov.communities.prsdb.webapp.forms.steps.DeregisterPropertyStepId
+import uk.gov.communities.prsdb.webapp.services.PropertyDeregistrationService
 import uk.gov.communities.prsdb.webapp.services.PropertyOwnershipService
-import uk.gov.communities.prsdb.webapp.services.PropertyRegistrationService
 import uk.gov.communities.prsdb.webapp.services.PropertyService
 import uk.gov.communities.prsdb.webapp.testHelpers.mockObjects.MockLandlordData
 import kotlin.test.assertEquals
@@ -28,17 +28,17 @@ import kotlin.test.assertEquals
 class DeregisterPropertyControllerTests(
     @Autowired val webContext: WebApplicationContext,
 ) : ControllerTest(webContext) {
-    @MockBean
+    @MockitoBean
     private lateinit var propertyDeregistrationJourneyFactory: PropertyDeregistrationJourneyFactory
 
-    @MockBean
+    @MockitoBean
     private lateinit var propertyOwnershipService: PropertyOwnershipService
 
-    @MockBean
+    @MockitoBean
     private lateinit var propertyService: PropertyService
 
-    @MockBean
-    private lateinit var propertyRegistrationService: PropertyRegistrationService
+    @MockitoBean
+    private lateinit var propertyDeregistrationService: PropertyDeregistrationService
 
     private val initialStepIdUrlSegment = PropertyDeregistrationJourney.initialStepId.urlPathSegment
 
@@ -118,7 +118,7 @@ class DeregisterPropertyControllerTests(
         val propertyOwnershipId = 1.toLong()
         val propertyId = 2.toLong()
         whenever(
-            propertyRegistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession(),
+            propertyDeregistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession(),
         ).thenReturn(mutableListOf(Pair(propertyOwnershipId, propertyId)))
 
         mvc
@@ -148,7 +148,7 @@ class DeregisterPropertyControllerTests(
     fun `getConfirmation returns 404 if the propertyOwnershipId is not in the list of deregistered propertyOwnershipIds in the session`() {
         val deregisteredPropertyEntities = mutableListOf(Pair(2.toLong(), 3.toLong()))
         val propertyOwnershipId = 1.toLong()
-        whenever(propertyRegistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession())
+        whenever(propertyDeregistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession())
             .thenReturn(deregisteredPropertyEntities)
 
         mvc
@@ -165,7 +165,7 @@ class DeregisterPropertyControllerTests(
         val propertyOwnership = MockLandlordData.createPropertyOwnership()
         val propertyOwnershipId = propertyOwnership.id
         whenever(
-            propertyRegistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession(),
+            propertyDeregistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession(),
         ).thenReturn(mutableListOf(Pair(propertyOwnershipId, propertyOwnership.property.id)))
         whenever(propertyOwnershipService.retrievePropertyOwnershipById(propertyOwnershipId)).thenReturn(propertyOwnership)
 
@@ -186,7 +186,7 @@ class DeregisterPropertyControllerTests(
         val property = propertyOwnership.property
 
         whenever(
-            propertyRegistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession(),
+            propertyDeregistrationService.getDeregisteredPropertyAndOwnershipIdsFromSession(),
         ).thenReturn(mutableListOf(Pair(propertyOwnershipId, propertyOwnership.property.id)))
         whenever(propertyService.retrievePropertyById(property.id)).thenReturn(property)
